Add TabEventKey and check to avoid useless api call, remove components
