+++ src/components/external/ApiSimilarShowTable.tsx
@@ -3,19 +3,18 @@ import { Table } from "react-bootstrap";
 import { ApiSimilarShow } from "../../models/external/ApiSimilarShow";
 import searchService from "../../services/searchService";
 import { errorToast } from "../../helpers/toasts";
+import TabEventKey from "../../models/internal/TabEventKey";
+import { TabProps } from "../../models/internal/TabProps";
 
-interface Props {
-    showId: number
-}
-
-export default function ApiSimilarShowTable({ showId }: Props) {
+export default function ApiSimilarShowTable({ showId, tabKey }: TabProps) {
     const [shows, setShows] = useState<ApiSimilarShow[]>([]);
 
     useEffect(() => {
         getSimilarsShows();
-    }, []);
+    }, [tabKey]);
 
     const getSimilarsShows = async () => {
+        if (tabKey !== TabEventKey.ApiSimilars || shows.length > 0) return
         const resp = await searchService.getSimilarByShowId(showId);
 
         if (resp.status === 200)
