+++ src/models/internal/TabEventKey.ts
@@ -1,0 +1,16 @@
+enum TabEventKey {
+    
+    Seasons = "seasons",
+
+    ApiSeasons = "api-seasons",
+
+    ApiImages = "api-images",
+
+    ApiSimilars = "api-similars",
+
+    ApiShowInfo = "api-show-infos",
+
+    ApiCharacters = "api-characters"
+}
+
+export default TabEventKey;
