+++ src/models/internal/TabProps.ts
@@ -1,0 +1,10 @@
+import TabEventKey from "./TabEventKey"
+
+interface TabProps {
+ 
+    showId: number
+
+    tabKey: TabEventKey
+}
+
+export type { TabProps };
