+++ src/views/discover/Discover.tsx
@@ -1,10 +1,14 @@
-import { Container, Tab, Tabs } from "react-bootstrap";
-import ApiDiscoverTitleRow from "../../components/external/ApiDiscoverTitleRow";
-import ApiDiscoverKindsRow from "../../components/external/ApiDiscoverKindsRow";
+import { Button, Col, Container, Form, Row, Tab, Tabs } from "react-bootstrap";
 import Navigation from "../../components/Navigation";
+import { ApiShowDetails } from "../../models/external/ApiShowDetails";
+import { useState, useEffect } from "react";
+import Loading from "../../components/Loading";
+import ApiShowCard from "../../components/external/ApiShowCard";
+import { errorToast } from "../../helpers/toasts";
+import searchService from "../../services/searchService";
+import { ApiShowKind } from "../../models/external/ApiShowKind";
 
 export default function Discover() {
-
     return (
         <Container className="mb-3">
             <Navigation url={'/discover'} />
@@ -23,3 +27,102 @@ export default function Discover() {
         </Container>
     );
 };
+
+function ApiDiscoverTitleRow() {
+    const [shows, setShows] = useState<ApiShowDetails[]>([]);
+    const [search, setSearch] = useState("");
+
+    useEffect(() => {
+        getShowsToDiscover();
+    }, [search]);
+
+    const getShowsToDiscover = async () => {
+        const resp = await searchService.discoverShows(search);
+
+        if (resp.status === 200)
+            setShows(await resp.json());
+        else
+            errorToast(await resp.json());
+    }
+
+    const onSubmit = async (e: any) => {
+        e.preventDefault();
+        setShows([]);
+        setSearch(e.target.titleSearch.value);
+    }
+
+    return (
+        <>
+            <Form className="mt-3" onSubmit={onSubmit}>
+                <Row>
+                    <Col>
+                        <Form.Group className="mb-3" controlId="titleSearch">
+                            <Form.Control type="text" placeholder="Titre de la série" required />
+                        </Form.Group>
+                    </Col>
+                    <Col>
+                        <Button variant="outline-dark" type="submit">
+                            <i className="bi-search"></i>
+                        </Button>
+                    </Col>
+                </Row>
+            </Form>
+
+            {shows.length > 0 ? <Row xs={2} md={3} lg={4}>
+                {shows.map(s => (
+                    <Col key={s.id} >
+                        <ApiShowCard show={s} />
+                    </Col>
+                ))}
+            </Row> : <Loading />}
+        </>
+    );
+}
+
+function ApiDiscoverKindsRow() {
+    const [kinds, setKinds] = useState<ApiShowKind[]>([]);
+    const [shows, setShows] = useState<ApiShowDetails[]>([]);
+    const [kind, setKind] = useState<string>("Comedy");
+
+    useEffect(() => {
+        getKinds();
+        getShowsByKind();
+    }, [kind]);
+
+    const getShowsByKind = async () => {
+        const resp = await searchService.getShowsByKind(kind);
+
+        if (resp.status === 200)
+            setShows(await resp.json());
+        else
+            errorToast(await resp.json());
+    }
+
+    const getKinds = async () => {
+
+        if (kinds.length > 0) return
+
+        const resp = await searchService.getKinds();
+
+        if (resp.status === 200)
+            setKinds(await resp.json());
+        else
+            errorToast(await resp.json());
+    }
+
+    return (
+        <>
+            <Form.Select onChange={(e => setKind(e.target.value))}>
+                {kinds.map(k => <option key={k.value} value={k.value}>{k.name}</option>)}
+            </Form.Select>
+
+            {shows.length > 0 ? <Row xs={2} md={3} lg={4} className="mt-3">
+                {shows.map(s => (
+                    <Col key={s.id} >
+                        <ApiShowCard show={s} />
+                    </Col>
+                ))}
+            </Row> : <Loading />}
+        </>
+    );
+}
