+++ src/views/discover/DiscoverDetails.tsx
@@ -2,21 +2,24 @@ import { useEffect, useState } from "react";
 import { useLocation, useNavigate, useParams } from "react-router-dom";
 import { ApiShowDetails } from "../../models/external/ApiShowDetails";
 import searchService from "../../services/searchService";
-import { Button, Container, Stack, Tab, Tabs, Image } from "react-bootstrap";
+import { Button, Container, Stack, Tab, Tabs, Image, Card, Col, Row } from "react-bootstrap";
 import Navigation from "../../components/Navigation";
 import Loading from "../../components/Loading";
 import { errorToast } from "../../helpers/toasts";
-import ApiCharactersRow from "../../components/external/ApiCharactersRow";
 import ApiShowInfos from "../../components/external/ApiShowInfos";
 import showService from "../../services/showService";
 import { getImageUrl } from "../../models/external/ApiShowImage";
 import ApiSimilarShowTable from "../../components/external/ApiSimilarShowTable";
+import TabEventKey from "../../models/internal/TabEventKey";
+import { ApiCharacterPreview } from "../../models/external/ApiCharacterPreview";
+import { TabProps } from "../../models/internal/TabProps";
 
 export default function DiscoverDetails() {
-    const { id } = useParams();
     const navigate = useNavigate();
+    const { id } = useParams();
     const { state } = useLocation();
     const [show, setShow] = useState<ApiShowDetails>();
+    const [key, setKey] = useState(TabEventKey.ApiShowInfo);
 
     useEffect(() => {
         if (state) {
@@ -62,20 +65,57 @@ export default function DiscoverDetails() {
                 </Stack>
 
                 <Tabs
-                    defaultActiveKey="infos"
+                    activeKey={key}
                     className="my-3"
+                    onSelect={(k) => setKey(k! as TabEventKey)}
                 >
-                    <Tab eventKey="infos" title="Infos">
+                    <Tab eventKey={TabEventKey.ApiShowInfo} title="Infos">
                         <ApiShowInfos show={show} />
                     </Tab>
-                    <Tab eventKey="characters" title="Acteurs">
-                        <ApiCharactersRow showId={show.id} />
+                    <Tab eventKey={TabEventKey.ApiCharacters} title="Acteurs">
+                        <ApiCharactersRow showId={show.id} tabKey={key} />
                     </Tab>
-                    <Tab eventKey="similar" title="Similaires">
-                        <ApiSimilarShowTable showId={show.id} />
+                    <Tab eventKey={TabEventKey.ApiSimilars} title="Similaires">
+                        <ApiSimilarShowTable showId={show.id} tabKey={key} />
                     </Tab>
                 </Tabs>
             </> : <Loading />}
         </Container>
     );
 };
+
+function ApiCharactersRow({ showId, tabKey }: TabProps) {
+    const [characters, setCharacters] = useState<ApiCharacterPreview[]>([]);
+
+    useEffect(() => {
+        getCharacters();
+    }, [tabKey]);
+
+    const getCharacters = async () => {
+        if (tabKey !== TabEventKey.ApiCharacters || characters.length > 0) return
+        const resp = await searchService.getCharactersByShowId(showId);
+
+        if (resp.status === 200)
+            setCharacters(await resp.json());
+        else
+            errorToast(await resp.json());
+    }
+
+    return (
+        <>
+            {characters.length > 0 ? <Row xs={2} md={3} lg={4} className="mt-4">
+                {characters.map(character => (
+                    <Col key={character.id} >
+                        <Card className="mt-2">
+                            <Card.Body>
+                                {character.picture && <Card.Img variant="top" src={character.picture} />}
+                                <Card.Title>{character.actor}</Card.Title>
+                                <Card.Subtitle>{character.name}</Card.Subtitle>
+                            </Card.Body>
+                        </Card>
+                    </Col>
+                ))}
+            </Row> : <Loading />}
+        </>
+    );
+}
